compress all cbr/cbz from a folder using CompressCbr.java
This script is for Unix environment.
It uses CompressCbr.java (that you can find in my Java repository)
It will compress all the cbr or cbz of the local folder using the CompressCbr.java
Unrar is required to extract cbr.
The output is a cbz (zip is used). The initial archive is not deleted.